Wrap BranchDataProviders in telemetry
