+++ src/api/createWrappedAzureResourcesExtensionApi.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { callWithTelemetryAndErrorHandlingSync, IActionContext } from '@microsoft/vscode-azext-utils';
 import { AzureResourcesApiInternal } from '../../hostapi.v2.internal';
+import { wrapFunctionsInTelemetry } from '../utils/wrapFunctionsInTelemetry';
 
 export function createWrappedAzureResourcesExtensionApi(api: AzureResourcesApiInternal, extensionId: string): AzureResourcesApiInternal {
 
@@ -32,33 +32,4 @@ export function createWrappedAzureResourcesExtensionApi(api: AzureResourcesApiIn
             }),
         }
     });
-}
-
-interface WrapFunctionsInTelemetryOptions {
-    /**
-     * Called before each function is executed. Intended for adding telemetry properties.
-     */
-    beforeHook?(context: IActionContext): void;
-    /**
-     * Optionally add a prefix to all function callbackIds.
-     */
-    callbackIdPrefix?: string;
-}
-
-function wrapFunctionsInTelemetry<TFunctions extends Record<string, (...args: unknown[]) => unknown>>(functions: TFunctions, options?: WrapFunctionsInTelemetryOptions): TFunctions {
-    const wrappedFunctions = {};
-
-    Object.entries(functions).forEach(([functionName, func]) => {
-        wrappedFunctions[functionName] = (...args: Parameters<typeof func>): ReturnType<typeof func> => {
-            return callWithTelemetryAndErrorHandlingSync((options?.callbackIdPrefix ?? '') + functionName, context => {
-                context.errorHandling.rethrow = true;
-                context.errorHandling.suppressDisplay = true;
-                context.errorHandling.suppressReportIssue = true;
-                options?.beforeHook?.(context);
-                return func(...args);
-            });
-        }
-    });
-
-    return wrappedFunctions as TFunctions;
 }
